[IMP] proxy_action : pre-commit stuff
